meisai_checker/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
python-docx>=1.1.2
PyMuPDF>=1.23.0
janome>=0.5.0
rapidfuzz>=3.9.0
streamlit>=1.30.0
